for bonus

+++ bonus/ft_lstsize_bonus.c
@@ -6,15 +6,15 @@
 /*   By: codespace <codespace@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/20 05:24:39 by codespace         #+#    #+#             */
-/*   Updated: 2025/07/20 08:33:24 by codespace        ###   ########.fr       */
+/*   Updated: 2025/07/21 07:21:49 by codespace        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../libft.h"
 
-size_t	ft_lstsize(t_list *lst)
+int	ft_lstsize(t_list *lst)
 {
-	size_t	count;
+	int	count;
 
 	if (!lst)
 		return (0);
